// src/language-code.enum.ts
export enum LanguageCode {
    Achinese = 'ace',
    Adangme = 'ada',
    Afrikaans = 'afr',
    Amharic = 'amh',
    Arabic = 'arb',
    AssyrianNeoAramaic = 'aii',
    Azerbaijani = 'azj', // North Azerbaijani (Latin)
    Balinese = 'ban',
    Belarusian = 'bel',
    Bemba = 'bem', // Bemba (Zambia)
    Bengali = 'ben',
    Bhojpuri = 'bho',
    Bini = 'bin',
    Tibetan = 'bod',
    Bosnian = 'bos', // Bosnian (Latin)
    Bulgarian = 'bul',
    Bulu = 'bum',
    Catalan = 'cat',
    Czech = 'ces',
    Chokwe = 'cjk',
    MandarinChinese = 'cmn',
    Danish = 'dan',
    German = 'deu',
    NortheasternDinka = 'dip',
    Dyula = 'dyu',
    Greek = 'ell',
    English = 'eng',
    Esperanto = 'epo',
    Ewe = 'ewe',
    French = 'fra',
    Pular = 'fuf',
    Fulfulde = 'fuv',
    Galician = 'glg',
    Gujarati = 'guj',
    Hausa = 'hau',
    Hebrew = 'heb',
    Hiligaynon = 'hil',
    Hindi = 'hin',
    QiandongMiao = 'hms',
    HmongNjua = 'hnj',
    Croatian = 'hrv',
    Hungarian = 'hun',
    Armenian = 'hye',
    Ibibio = 'ibb',
    Ido = 'ido',
    Indonesian = 'ind',
    Ilocano = 'ilo',
    Italian = 'ita',
    Javanese = 'jav', // Javanese (Latin)
    Japanese = 'jpn',
    Kazakh = 'kaz',
    Makonde = 'kde',
    Khmer = 'khm',
    Kinyarwanda = 'kin',
    Kirghiz = 'kir',
    Kimbundu = 'kmb',
    KomiPermyak = 'koi',
    Korean = 'kor',
    Kwangali = 'kwn',
    Lao = 'lao',
    Lithuanian = 'lit',
    Lingala = 'lin',
    Latvian = 'lvs',
    Magahi = 'mag',
    Maithili = 'mai',
    Malayalam = 'mal',
    Marathi = 'mar',
    Macedonian = 'mkd',
    Malagasy = 'mlg',
    Malay = 'mly',
    Mooré = 'mos',
    Burmese = 'mya',
    Dutch = 'nld',
    NorwegianBokmal = 'nob',
    Nepali = 'npi',
    Pedi = 'nso', // Sesotho sa Leboa
    Nyanja = 'nya',
    Occitan = 'oci',
    Oriya = 'ori',
    Panjabi = 'pan',
    IranianPersian = 'pes',
    MalagasyPlateau = 'plt',
    Polish = 'pol',
    Portuguese = 'por',
    Dari = 'prs',
    ChimborazoHighlandQuichua = 'qug',
    CuscoQuechua = 'quz',
    Rundi = 'run',
    Russian = 'rus',
    Sango = 'sag',
    Sinhala = 'sin',
    Slovak = 'slk',
    Slovenian = 'slv',
    NorthernSami = 'sme',
    Shona = 'sna',
    Sotho = 'sot',
    Spanish = 'spa',
    Sardinian = 'srd',
    Serbian = 'srp', // Serbian (Cyrillic)
    Swati = 'ssw',
    Sundanese = 'sun',
    Swedish = 'swe',
    Tamil = 'tam',
    Tatar = 'tat',
    Telugu = 'tel',
    Tajik = 'tgk',
    Thai = 'tha',
    Tigrinya = 'tir',
    Turkmen = 'tuk',
    Turkish = 'tur',
    Ukrainian = 'ukr',
    Urdu = 'urd',
    Uzbek = 'uzn', // Northern Uzbek (Latin)
    Venda = 'ven',
    Vietnamese = 'vie',
    Wolof = 'wol',
    Xhosa = 'xho',
    Yoruba = 'yor',
    Cantonese = 'yue',
    Zulu = 'zul',
}