// src/language.enum.ts
export enum Language {
    ace = 'Achinese',
    ada = 'Adangme',
    afr = 'Afrikaans',
    amh = 'Amharic',
    arb = 'العربية',
    aii = 'Assyrian Neo-Aramaic',
    azj = 'Azərbaycan dili', // North Azerbaijani (Latin)
    ban = 'Basa Bali', // Balinese
    bel = 'Беларуская',
    bem = 'Ichibemba', // Bemba (Zambia)
    ben = 'বাংলা',
    bho = 'भोजपुरी',
    bin = 'Ẹ̀dó',
    bod = 'བོད་ཡིག', // Tibetan
    bos = 'Bosanski', // Bosnian (Latin)
    bul = 'Български',
    bum = 'Bulu',
    cat = 'Català',
    ces = 'Čeština',
    cjk = 'Cokwe',
    cmn = '普通话', // Mandarin Chinese
    dan = 'Dansk',
    deu = 'Deutsch',
    dip = 'Thuɔŋjäŋ', // Northeastern Dinka
    dyu = 'Jula', // Dyula
    ell = 'Ελληνικά',
    eng = 'English',
    epo = 'Esperanto',
    ewe = 'Eʋegbe',
    fra = 'Français',
    fuf = 'Pulaar', // Pular
    fuv = 'Fulfulde',
    glg = 'Galego',
    guj = 'ગુજરાતી',
    hau = 'Hausa',
    heb = 'עברית',
    hil = 'Ilonggo',
    hin = 'हिन्दी',
    hms = 'Qiandong Miao',
    hnj = 'Hmong Njua',
    hrv = 'Hrvatski',
    hun = 'Magyar',
    hye = 'Հայերեն',
    ibb = 'Ibibio',
    ido = 'Ido',
    ind = 'Bahasa Indonesia',
    ilo = 'Ilocano',
    ita = 'Italiano',
    jav = 'Basa Jawa', // Javanese (Latin)
    jpn = '日本語',
    kaz = 'Қазақ тілі',
    kde = 'Chimakonde',
    khm = 'ខ្មែរ', // Khmer
    kin = 'Ikinyarwanda',
    kir = 'Кыргызча',
    kmb = 'Kimbundu',
    koi = 'Коми-пермяк', // Komi-Permyak
    kor = '한국어',
    kwn = 'Kwangali',
    lao = 'ລາວ',
    lit = 'Lietuvių',
    lin = 'Lingála',
    lvs = 'Latviešu',
    mag = 'मगही',
    mai = 'मैथिली',
    mal = 'മലയാളം',
    mar = 'मराठी',
    mkd = 'Македонски',
    mlg = 'Malagasy',
    mly = 'Bahasa Melayu',
    mos = 'Mooré',
    mya = 'ဗမာစာ', // Burmese
    nld = 'Nederlands',
    nob = 'Norsk bokmål',
    npi = 'नेपाली',
    nso = 'Sesotho sa Leboa', // Pedi
    nya = 'Chinyanja',
    oci = 'Occitan',
    ori = 'ଓଡ଼ିଆ',
    pan = 'ਪੰਜਾਬੀ',
    pes = 'فارسی', // Iranian Persian
    plt = 'Malagasy Plateau',
    pol = 'Polski',
    por = 'Português',
    prs = 'دری', // Dari
    qug = 'Runa shimi', // Chimborazo Highland Quichua
    quz = 'Runa Simi', // Cusco Quechua
    run = 'Kirundi',
    rus = 'Русский',
    sag = 'Sango',
    sin = 'සිංහල',
    slk = 'Slovenčina',
    slv = 'Slovenščina',
    sme = 'Davvisámegiella',
    sna = 'ChiShona',
    sot = 'Sesotho',
    spa = 'Español',
    srd = 'Sardu',
    srp = 'Српски', // Serbian (Cyrillic)
    ssw = 'SiSwati',
    sun = 'Basa Sunda',
    swe = 'Svenska',
    tam = 'தமிழ்',
    tat = 'Татарча', // Tatar
    tel = 'తెలుగు',
    tgk = 'Тоҷикӣ',
    tha = 'ไทย',
    tir = 'ትግርኛ',
    tuk = 'Türkmen',
    tur = 'Türkçe',
    ukr = 'Українська',
    urd = 'اُردُو',
    uzn = 'Oʻzbek', // Northern Uzbek (Latin)
    ven = 'Tshivenḓa',
    vie = 'Tiếng Việt',
    wol = 'Wolof',
    xho = 'isiXhosa',
    yor = 'Èdè Yorùbá',
    yue = '粤语',
    zul = 'isiZulu',
}